Provide "self" argument instead of <nixpkgs>
If one custom package "A" depended on another "B", then A's definition
would use "(import <nixpkgs> {}).B" to access it. That was fine when we
could rely on packageOverrides to be picked up from ~/.nixpkgs, but in
environments like Hydra that's not an option.

Instead, we now pass in the overridden packages explicitly, as a "self"
argument. For consistency with Nix convention, the non-overridden
packages, previously called "pkgs", are now called "super". Hence the
definition of A would now look like:

self: super:
...
self.B
...