BREAKING CHANGE: Don't allow latestGit when 'stable' is true
Fetching the latest version of some repo defeats the point of using a stable
version of nix-config, hence we forbid the use of latestGit in these cases; with
two exceptions:

 - latestNixCfg gives access to an unstable nix-config, from within a stable one
 - runCabal2nix has a workaround for dodgy git handling in cabal2nix, such that
   we check out a given repo first, then call cabal2nix on that directory. Since
   we have no commit info for these, we allow latestGit, with a warning (since
   such usages should really be replaced by an explicit call to latestGit, with
   a proper fallback, etc.)

Notice that this causes a lot of derivations to change, and it will cause
external projects which call into this repo break as well. Going forward, that
should be less of a problem, since they can use an *actually* stable version.