Patch space2ctrl to avoid triggering a space at start
The space2ctrl code triggers a keyup/keydown combo for keycode 255 (which it
remaps to space) during its 'start' method. This causes a space to be sent to
whichever application is focused. In the best case this is an annoyance, but it
seems pretty bad if we'll be restarting space2ctrl at arbitrary
times (e.g. when plugging in a keyboard).

I've added a patch to the derivation which comments-out the offending code. Note
that the code has the following comment:

    // TODO: document why the following event is needed

So presumably it is needed for some reason, but my quick testing of the patched
version shows that it seems to work fine without it. In any case, I'd call this
event a bug, regardless of whether it works around some other undocumented bug.