Simplify pandocPkgs
The biggest improvement is that we now use pinned, pre-generated versions of
everything. We don't require running cabal or cabal2nix at eval time, which is a
big win since those are hefty dependencies which we don't want to be building
unless it's build time.

Another improvement is that we're taking more things from nixpkgs's binary
cache, which is a big win as long as those cached packages remain
available (can't wait until an IPFS store becomes practical).

The downsideis that it may be more painful to upgrade or swap out packages in
the future. This is because we just-so-happen to be taking most things from the
nixpkgs haskellPackages sets with only minimal alterations (hence the binary
cache hits). Yet there's nothing to say that a small bump in version won't cause
a cascade of conflicts, which we'd have to resolve using cabal and cabal2nix
again (although we could probably pre-generate the set, rather than relying on
import-from-derivation again).