Move a bunch of hasBinary checks to 'tests' rather than 'withDeps'
It's nice to have packages depend on their test suite passing, but one downside
of 'withDeps' is that the resulting derivation is distinct, in the eyes of Nix,
and hence needs to be built *as well as* the input derivation. For simple
wrappers, like 'buildEnv', this is fine. For substantial builds it's not very
nice though.

Lots of these packages were also just copying their derivation to the 'tests'
attribute anyway, which is less semantically meaningful than putting their
actual tests in there.