Remove tasty Haskell overrides
I think these were needed by some packages that weren't compatible with the
default Haskell package set of an old NixOS version, but they should go for a
few reasons:

 - NixOS has been upgraded since then, and it's unlikely that these hard-coded
   versions will fix the new package sets too
 - It's not clear which packages these were meant to fix. It would be far better
   to put the overrides into those package's definitions, rather than altering
   the global set.
 - It's unlikely that many different packages can co-exist with the same Haskell
   package set; rather than attempting this (which the NixOS authors have
   already done!) we should just check if a package works, and if not we should
   override the package set it uses *locally*.
 - By altering the global package set, we get a lot of binary cache misses; even
   for packages which don't need this particular version of tasty.
 - We should strive to treat Haskell packages like libraries, i.e. accept
   any dpendencies we're given and see what happens, just in case some other
   library needs certain dependency versions. We should only enforce particular
   versions (like these tasty ones) for *applications*, which we can 'lift' out
   of the 'haskellPackages' set into the nixpkgs top-level.

For an example of this last point, we can imagine that e.g. brittany requires
some specific version of, say, tasty.

The worst approach would be to override tasty to that version in the global
haskellPackages set; since that would force everything to use that version.

The second-worst approach would be to override haskellPackages.brittany to use
that version of tasty. This is a local change, but it may cause inconsistencies
if the brittany package is used as a dependency of something else.

The best approach is to make a top-level 'brittany' package, which does
something like '(haskellPackages.override { tasty = ...; }).brittany'.

The downside of this last approach is that we might end up duplicating this
override for several packages. The correct solution to that problem is to factor
out either the overridden haskellPackages set or the set of tasty overrides into
a helper function which those packages can all use. For an example of this see
the 'panPkgs' helper.