Take tip-lib patching package from haskell context
We were taking the BNFC package, used for patching the tip source, from the
'haskellPackages' set, but that may break depending on the nixpkgs version, and
is hard to override.

Instead, we now take it from whichever haskell package set we're building
tip-lib against; this allows the usual haskell package set overriding to be used
for BNFC too.