Bump jsbeautifier
The old version's test suite started failing when upgrading to NixOS 19.09, due
to a missing input .js file for testing "myobfuscator". The .js file didn't
appear in the source archive, Googling for the filename showed that it
should/did exist. The sha256 hadn't changed from when it was working, so that's
weird.

Rather than try to debug it, I instead tried upgrading. The latest version,
1.10.3, requires a newer version of pythonPackages.six than Nixpkgs has, so this
commit brings us up to 1.10.2 instead.