Check that haskell-nix build plans are cached
Running cabal to calculate a build plan can take several seconds, and this is
done once per Haskell project. We cache the results to avoid needing to
recalculate them every time, and check to see whether haskell-nix occurrences
are accompanied by a reference to the cache. It's also important to update the
cache whenever changes are made.