getDeps: use callCabal2nixWithPlan (WIP)
This is an easier way to define Haskell packages, and doesn't require
manually-updating hashes, etc.

Unfortunately GetDeps depends on atto-lisp, which doesn't work with GHC > 8.6.
Annoyingly, the package itself sets an upper-bound on base < 5; but that was
brought down by Hackage revisions. Will probably require a patch/fork of
atto-lisp?