skulpture: fix Qt5 build
Not entirely sure what the problem was, but running QT_DEBUG_PLUGINS=1 qt5ct
was showing:

    "Failed to extract plugin meta data from '/nix/store/6jygkb9ffgsicm5qyc222qgy1ah6pffi-skulpture-qt5/lib/qt-5.15.12/plugins/styles/skulpture.so'"
         not a plugin

This might have been due to building the config gui, and accidentally
overwriting the desired library?

In any case, this change replaces the postInstall shell commands which tried to
find the libraries and put them in place; instead using a preConfigure command
to set CMake options that will do it "properly".

Running qt5ct with the resulting path in QT_PLUGIN_PATH seems to work!