Add multibase-cli
The repo is mostly for the multibase library, but we only care about the
'multibase' binary defined in its cli/ subfolder.

The upstream version is rather awkward, since it only takes data in
arguments (which isn't binary-safe), and prints extra output like "Result: "
which makes it harder to extract the desired data. My fork removes the latter,
and accepts data on stdin.