git-on-ipfs: Use newScope to allow overrides
Specifically, the '.override' function for each thing in packages/ will be for
the function defined by its default.nix file; NOT for the return value of such a
call. Hence if we do e.g. '{ callPackage }: callPackage foo {}' then we get a
'.override' for replacing that '{ callPackage }' argument set; not for the 'foo'
function.

To work around this, we use ', ... }@args' to accept arbitrary extra
arguments (including whatever our git-on-ipfs function might want), and we use
'newScope args git-on-ipfs {}' so that the given 'args' will override the attrs
used to select argument values for git-on-ipfs. It's a little indirect, and
won't give us error messages if we misspell an argument name, but it's a simple
way to allow overriding.